refactor(parser): rename and reorganize methods for clarity and modularity
- Renamed `CLOSE_FOR_KEY` to `FOR_CLOSE_KEY` across lexer, parser, and tests for consistency.
- Reorganized parser methods: extracted individual parsing logic into separate functions (`parsSet`, `parseExpression`, etc.).
- Updated `PareserElementFactory` to `ParserElementFactory` for corrected naming.
- Enhanced test data to align with updated token and method names.
