+++ src/test/resources/testdata/for/CompleteFor.txt
@@ -9,5 +9,5 @@ Vento File(0,34)
   PsiWhiteSpace('\n')(23,24)
   VentoPsiElement: VENTO_FOR_ELEMENT(24,34)
     PsiElement(VENTO_FOR_START_TOKEN)('{{')(24,26)
-    PsiElement(VENTO_CLOSE_FOR_KEY_TOKEN)(' /for')(26,31)
+    PsiElement(VENTO_FOR_CLOSE_KEY_TOKEN)(' /for')(26,31)
     PsiElement(VENTO_FOR_END_TOKEN)(' }}')(31,34)
