+++ src/test/resources/testdata/for/CompleteForBrokenObject.txt
@@ -14,5 +14,5 @@ Vento File(0,41)
   PsiWhiteSpace('\n')(30,31)
   VentoPsiElement: VENTO_FOR_ELEMENT(31,41)
     PsiElement(VENTO_FOR_START_TOKEN)('{{')(31,33)
-    PsiElement(VENTO_CLOSE_FOR_KEY_TOKEN)(' /for')(33,38)
+    PsiElement(VENTO_FOR_CLOSE_KEY_TOKEN)(' /for')(33,38)
     PsiElement(VENTO_FOR_END_TOKEN)(' }}')(38,41)
