+++ src/test/resources/testdata/for/CompleteForObject.txt
@@ -14,5 +14,5 @@ Vento File(0,42)
   PsiWhiteSpace('\n')(31,32)
   VentoPsiElement: VENTO_FOR_ELEMENT(32,42)
     PsiElement(VENTO_FOR_START_TOKEN)('{{')(32,34)
-    PsiElement(VENTO_CLOSE_FOR_KEY_TOKEN)(' /for')(34,39)
+    PsiElement(VENTO_FOR_CLOSE_KEY_TOKEN)(' /for')(34,39)
     PsiElement(VENTO_FOR_END_TOKEN)(' }}')(39,42)
